[REF] Change type of the offset to off_t
When accessing storage we need to be able to address even large storages, so
the offset must be of an off_t type.

Fixes:
SEOS-1472

+++ components/Tests/src/test_OS_FileSystemFile.c
@@ -81,7 +81,7 @@ static void
 test_OS_FileSystemFile_getSize(
     OS_FileSystem_Handle_t hFs)
 {
-    size_t size;
+    off_t size;
 
     TEST_START();
